+++ src/components/TextField.tsx
@@ -2,18 +2,6 @@ import React, { FunctionComponent } from "react";
 import { InputProps } from "../types/InputProps";
 import useInput from "../hooks/useInput";
 
-export const min = (length: number) => (value: string | number) => {
-  if (String(value).length < length)
-    return `최소 ${length}글자 이상 입력해주세요`;
-  else return undefined;
-};
-
-export const max = (length: number) => (value: string | number) => {
-  if (String(value).length > length)
-    return `최대 ${length}글자 이하로 입력해주세요`;
-  else return undefined;
-};
-
 const TextField: FunctionComponent<InputProps> = ({
   source,
   label,
