+++ src/validation/index.ts
@@ -1,0 +1,16 @@
+export const required = (message: string) => (value?: string) => {
+    if (value === undefined || value === null) return message;
+    else return undefined;
+};
+
+export const min = (length: number) => (value: string | number) => {
+    if (String(value).length < length)
+        return `최소 ${length}글자 이상 입력해주세요`;
+    else return undefined;
+};
+
+export const max = (length: number) => (value: string | number) => {
+    if (String(value).length > length)
+        return `최대 ${length}글자 이하로 입력해주세요`;
+    else return undefined;
+};
